MB-37313: Add magma function to get valueSize from metadata
As part of CBSS-332, adding metadata to seq index, we need a way to get
the value size from the metadata.

For magma, we encode the value with both metadata and value. We also
store the value in the key index when the value is small. When
traversing the seq index, we need to know if the value is stored in the
key index by looking at the value size from the metadata. We need a
function to get that value size.

Change-Id: Ic21133f5c64f892efaaaa8f36a3dc11aed365e9c
Reviewed-on: http://review.couchbase.org/119572
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>