MB-37917: Reset initialDiskSnapshot flag on transition to active
Currently we will reset the initialDiskSnapshot flag of a vBucket when
we transition from replica to non-replica. The intent of this change
was to ensure that if we transition to active without having received
a full snapshot then we would allow consumers to stream from us. This
does not consider that we may receive our initial disk snapshot whilst
we are a pending vBucket though. We should instead reset this flag
whenever we transition to active as we are now the source of truth and
should allow streaming from this vBucket.

Change-Id: I0b2ad8467e70575567eeffe0a25731281977b8e3
Reviewed-on: http://review.couchbase.org/122261
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>